Reenable line numbers in transcriptions, improve transcription display
We're also still optimizing how the text shows up on the server for SEO -- this should mean that inscription pages are easier to find in Google etc.

The downside is that it causes what look like flashes of unstyled content. Really, we're processing the raw edition XML in the browser and replacing the bare text with the renderable "TEI HTML" from the CETEIcean library.

It should eventually be possible to pass CETEIcean a DOM-like object so that we can fully render things on the server, but that's out of scope right now.